requirements.txt
flask
youtube_transcript_api
python-dotenv
google-generativeai
google-cloud-storage
requests
logging
